Add new Jekyll-based Markdown index page
This removes the original HTML file and replaces it with a Markdown file that
generates a very similar output. It also gives us a lot of flexibility going
forward to write new pages in Markdown and easily change to an alternative
theme.

This change also removes a copy of Slate from `third_party` as we no longer need
it, since we now download it via Ruby Gems.

Signed-off-by: Misha Brukman <mbrukman@google.com>