Adds a shield for easily seeing the website status
This badge shows at a glance whether the website is up or down; it could be down
for any number of reasons, e.g., a DNS misconfiguration on the part of the
domain owner or GitHub's hosting, or inability to reach the GitHub servers
hosting the website.

This visual indicator will allow us to more quickly notice the issue and fix it.

Signed-off-by: Misha Brukman <mbrukman@google.com>